+++ src/loaders.py
@@ -1,0 +1,7 @@
+import yaml
+
+
+def load_config(config_path):
+    with open(config_path, "r") as file:
+        config = yaml.safe_load(file)
+    return config
