--- src/model.py
@@ -1,27 +1,0 @@
-import torch.nn as nn
-
-
-class DQN(nn.Module):
-    def __init__(self, input_dim, action_dim):
-        super(DQN, self).__init__()
-        # Convolutional layers based on the Atari paper
-        self.conv_layers = nn.Sequential(
-            nn.Conv2d(input_dim[0], 32, kernel_size=8, stride=4),
-            nn.ReLU(),
-            nn.Conv2d(32, 64, kernel_size=4, stride=2),
-            nn.ReLU(),
-            nn.Conv2d(64, 64, kernel_size=3, stride=1),
-            nn.ReLU(),
-        )
-
-        # Fully connected layers
-        self.fc_layers = nn.Sequential(
-            nn.Flatten(),
-            nn.Linear(64 * 7 * 7, 512),  # Adjust based on input dimensions
-            nn.ReLU(),
-            nn.Linear(512, action_dim),
-        )
-
-    def forward(self, x):
-        x = self.conv_layers(x)
-        return self.fc_layers(x)
